// src/generated-api/models/CommonResponseAdminDiarySdohResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * CareDiary API
 * 돌봄일기 서비스
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { AdminDiarySdohResponse } from './AdminDiarySdohResponse';
import {
    AdminDiarySdohResponseFromJSON,
    AdminDiarySdohResponseFromJSONTyped,
    AdminDiarySdohResponseToJSON,
    AdminDiarySdohResponseToJSONTyped,
} from './AdminDiarySdohResponse';

/**
 * 
 * @export
 * @interface CommonResponseAdminDiarySdohResponse
 */
export interface CommonResponseAdminDiarySdohResponse {
    /**
     * 
     * @type {number}
     * @memberof CommonResponseAdminDiarySdohResponse
     */
    status: number;
    /**
     * 
     * @type {string}
     * @memberof CommonResponseAdminDiarySdohResponse
     */
    code: string;
    /**
     * 
     * @type {string}
     * @memberof CommonResponseAdminDiarySdohResponse
     */
    message: string;
    /**
     * 
     * @type {AdminDiarySdohResponse}
     * @memberof CommonResponseAdminDiarySdohResponse
     */
    data?: AdminDiarySdohResponse;
}

/**
 * Check if a given object implements the CommonResponseAdminDiarySdohResponse interface.
 */
export function instanceOfCommonResponseAdminDiarySdohResponse(value: object): value is CommonResponseAdminDiarySdohResponse {
    if (!('status' in value) || value['status'] === undefined) return false;
    if (!('code' in value) || value['code'] === undefined) return false;
    if (!('message' in value) || value['message'] === undefined) return false;
    return true;
}

export function CommonResponseAdminDiarySdohResponseFromJSON(json: any): CommonResponseAdminDiarySdohResponse {
    return CommonResponseAdminDiarySdohResponseFromJSONTyped(json, false);
}

export function CommonResponseAdminDiarySdohResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): CommonResponseAdminDiarySdohResponse {
    if (json == null) {
        return json;
    }
    return {
        
        'status': json['status'],
        'code': json['code'],
        'message': json['message'],
        'data': json['data'] == null ? undefined : AdminDiarySdohResponseFromJSON(json['data']),
    };
}

export function CommonResponseAdminDiarySdohResponseToJSON(json: any): CommonResponseAdminDiarySdohResponse {
    return CommonResponseAdminDiarySdohResponseToJSONTyped(json, false);
}

export function CommonResponseAdminDiarySdohResponseToJSONTyped(value?: CommonResponseAdminDiarySdohResponse | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'status': value['status'],
        'code': value['code'],
        'message': value['message'],
        'data': AdminDiarySdohResponseToJSON(value['data']),
    };
}